refactor: clarifying purpose of test scripts
