+++ test/test_tutorial_evaluatingbinaryclassification.py
@@ -7,7 +7,7 @@ Upon recommendation from the AIF360 development team
 '''
 
 from .notebook_tester import validate_notebook, list_warnings
-from .__test_utils import is_test_environment
+from .__testing_utilities import is_test_environment
 import os
 import sys
 import warnings
